Add testing and move directories logic
