+++ package.json
@@ -4,11 +4,13 @@
   "description": "Hola Nico, morirás de noche",
   "main": "dist/server.js",
   "scripts": {
-    "start": "node .",
-    "build": "npx tsc",
-    "dev": "nodemon server.js",
+    "serve": "node .",
+    "build": "npm run clean && npx tsc",
+    "clean": "rm -r ./dist/*",
+    "dev:watch": "nodemon server.js",
     "dev:ts": "ts-node ./src/server.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest",
+    "test:watch":"jest --watch"
   },
   "keywords": [],
   "author": "",
@@ -20,6 +22,7 @@
   "devDependencies": {
     "@types/body-parser": "^1.17.0",
     "@types/express": "^4.17.0",
+    "@types/jest": "^24.0.15",
     "@types/node": "^12.0.12",
     "@typescript-eslint/eslint-plugin": "^1.11.0",
     "@typescript-eslint/parser": "^1.11.0",
@@ -27,6 +30,7 @@
     "eslint-config-prettier": "^6.0.0",
     "eslint-plugin-prettier": "^3.1.0",
     "express-generator": "^4.16.1",
+    "jest": "^24.8.0",
     "nodemon": "^1.19.1",
     "prettier": "^1.18.2",
     "ts-node": "^8.3.0"
