+++ src/api/controllers/__test__/gamesController.test.ts
@@ -1,0 +1,5 @@
+describe('Games testing', () => {
+    it('Throw dice',() => {
+        expect(1).toBe(1)
+    })
+})
