+++ src/api/controllers/__test__/moviesController.test.ts
@@ -1,0 +1,5 @@
+describe('Movies testing', () => {
+    it('Movies test types',() => {
+        expect(1).toBe(1)
+    })
+})
