+++ src/api/controllers/__test__/usersController.test.ts
@@ -1,0 +1,5 @@
+describe('Users testing', () => {
+    it('User check datatype',() => {
+        expect(1).toBe(1)
+    })
+})
