--- src/api/controllers/index.ts
@@ -1,3 +1,0 @@
-export * from './gamesController'
-export * from './moviesController'
-export * from './usersController'
