+++ src/api/routers/index.ts
@@ -1,0 +1,3 @@
+export { gamesRouter } from './gamesRouter'
+export { moviesRouter } from './moviesRouter'
+export { usersRouter } from './usersRouter'
